requirements.txt
###### Requirements withtout version specifiers ######

discord.py
python-dotenv

###### Requirements with version specifiers ######